fixed docker build after eslint changes
[issue]
eslint was runing out of memory within a docker build.

[cause]
either eslint changed its behavior or
the changed way we invoke eslint in the package.json caused
eslint to scann the whole filesystem (/)
Or it ways there all along and I just noticed it now xD

[fix]
put the source into a dedicated folder.